First kv integration test
